Android: tools: dynamically link to libsamsung-ipc
There is no point of bundling a copy of libsamsung-ipc in the
tools.

It's also counter-intuitive as developers do expect the tools
to be able to test the same libsamsung-ipc that is used by
the system and not a separate copy.

Without that we also have some compilation issue with ipc-modem.
  external/libsamsung-ipc/samsung-ipc/rfs.c:55: error: undefined reference to 'MD5_Init'
  external/libsamsung-ipc/samsung-ipc/rfs.c:56: error: undefined reference to 'MD5_Update'
  external/libsamsung-ipc/samsung-ipc/rfs.c:57: error: undefined reference to 'MD5_Update'
  external/libsamsung-ipc/samsung-ipc/rfs.c:58: error: undefined reference to 'MD5_Final'

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>