modem: rename xmm626_sec_modem to xmm626_kernel_smdk4412
The XMM626 is the modem model, and SEC is a reference to
Samsung Electronics.

However, in practice the xmm626_sec_modem file and
functions are tied to a specific kernel interface which is
not upstream.

In CyanogenMod, LineageOS and Replicant (up to the 6.0
version), most of the devices supported by libsamsung-ipc
use a kernel named android_kernel_samsung_smdk4412, which
implements this interface.

So it is more clear to use a name that reflects that
instead of using the generic xmm626_sec_modem name.

Even if SMDK4412 is also a Samsung evaluation board for the
Exynos 4412 System on a chip, the kernel name is probably
way more known to people working on supporting devices with
an XMM626 modem in libsamsung-ipc.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>