samsung-ipc: utils: switch to Linux code style
This contains no functional changes.

The commented out code if any has also been preserved as it
could be relevant to the understanding of the code.

It will be kept until we understand why the code has been
commented. In that case we either need to remove it completely
or to replace it by a comment explaining why not having that
code was necessary.

For __FUNCTION__, according to checkpatch.pl, it's gcc
specific:
    WARNING: __func__ should be used instead of gcc specific __FUNCTION__
However __func__ is part of the C18 standard[1]. In the
"6.4.2.2 Predefined identifiers" section, we have:
    "The identifier __func__ shall be implicitly declared by
     the translator as if, immediately following the opening
     brace of each function definition, the declaration
         static const char __func__[] = "function-name";
    "

Using symbolic permissions for file mode triggered a
checkpatch.pl warning:
    WARNING: Symbolic permissions
    'S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH' are not
    preferred. Consider using octal permissions '0644'.

The rationale behind that is that octal is generally easier
to understand than the symbolic relresentation. For more
information, see the follwoing Linux commit:
f90774e1fd27 checkpatch: look for symbolic permissions and suggest octal instead

References:
-----------
[1]The standard doesn't seem to be available for free, but the draft
   can be downloaded from the following URL:
   https://web.archive.org/web/20181230041359if_/http://www.open-std.org/jtc1/sc22/wg14/www/abq/c17_updated_proposed_fdis.pdf

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Reviewed-by: Joonas Kylmälä <joonas.kylmala@iki.fi>