Add a tool to help check the conversion to Linux code style
When sending a patch for Linux, and running checkpatch.pl,
you typically only need to check the patch with it.

However since we're doing an initial conversion to the
Linux code style, we also need to check all the content of
the files as well, as otherwise the parts that are not
being touched by the patches won't be checked at all.

We also need to get it right, as otherwise running
checkpatch.pl on subsequent patches could trigger some
errors that would not come from the patches being checked,
but that would instead come from previous code style
errors, that were left unfixed.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Reviewed-by: Joonas Kylmälä <joonas.kylmala@iki.fi>