check_code_style_conversion.py: also check files that are not the revision range
This checks the files that aren't modified by any of the
commits that are in the revision range.

This is not completely failsafe as checkpatch.pl tend to
check better what's touched by a patch, in the case of
modifications that do not touch C code:
- Checking for scripts/spelling.txt with checkpatch.pl -f
  doesn't raise any errors or warnings.
- Checking the patch that adds scripts/spelling.txt
  raise a lot of warnings.

However checkpatch.pl seem to find issues in C code regardless
of whether the code is checked as part of individual files
or as part of a patch.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Reviewed-by: Joonas Kylmälä <joonas.kylmala@iki.fi>