Android.mk: remove deprecated LOCAL_COPY_HEADERS
The Android "Build System Best Practices" document[1] states:
    "LOCAL_COPY_HEADERS is deprecated. Soong modules cannot use these
     headers, and when the VNDK is enabled, System modules in Make cannot
     declare or use them either."
and:
    "Don't use non-local paths in your LOCAL_EXPORT_C_INCLUDE_DIRS, [...].
     Non-local exported include dirs are not supported in Soong."

And having both VNDK enabled and LOCAL_COPY_HEADERS does prevent the build
on Replicant 10.

LOCAL_COPY_HEADERS was kept for compatibility reasons in the fear that
not removing it would potentially break older Replicant versions like
Replicant 4.2 which are still needed in order to test libsamsung-ipc
on devices that are not supported anymore by subsequent Replicant versions.

However since:
- LOCAL_EXPORT_C_INCLUDE_DIRS is also used in Replicant 4.2 and 6.0 as
  well
- all devices currently supported by libsamsung-ipc are supported by one of
  these Replicant versions[2]
it should be safe to remove all usage of LOCAL_COPY_HEADERS in Android.

This change was tested at build time with the following configurations:
- Replicant 10.0 with the i9300 target
- Replicant  6.0 with the i9300 target
- Replicant  4.2 with the galaxysmtd target

References:
-----------
[1]https://android.googlesource.com/platform/build/soong/+/master/docs/best_practices.md#headers
[2]With the exception of the Galaxy Tab which was never supported by
   Replicant.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Joonas Kylmälä <joonas.kylmala@iki.fi>