Android.mk: add nv_data-md5
The nv_data-md5 tool can be used to create the nv_data.md5 or .nv_data.md5
files.

This tool is already integrated in the Autotools build system, so it's a
good idea to also add it to the Android.mk to enable to build it in Android
as well if needed.

As libsamsung-ipc is not packaged yet in most common GNU/Linux
distributions, it might be useful to add it in the resulting Android
images, but it's up to the Android distributions using libsamsung-ipc
to decide that, for instance by adding nv_data-md5 to PRODUCT_PACKAGES
in the repositories of various devices.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>