samsung-ipc: utils.c: file_data_{read,write}: report open error
Mounting the EFS on your local computer can result in user id
and permissions mismatch because the /etc/fstab doesn't always
match the user ids used by Android.

For instance here's the GT-N7000 EFS on my laptop:
    $ ls -l [...]/nv_data.bin
    -rwx------ 1 1001 1001 2097152  1 janv.  2000 [...]/nv_data.bin

When using nv_data-md5 on it we have:
    $ ./tools/nv_data-md5 [...]/nv_data.bin
    [ipc] file_data_read: Error: fd: -1
    [ipc] ipc_nv_data_md5_calculate failed: data is NULL
    Calculating nv_data backup md5 failed

The error was too cryptic, and I ended up having to dig into the source
code to understand what was going on.

With this patch we now have an error message that is easier to understand:
    $ ./tools/nv_data-md5 [...]/nv_data.bin
    [ipc] file_data_read open failed with error 13: Permission denied
    [ipc] ipc_nv_data_md5_calculate failed: data is NULL
    Calculating nv_data backup md5 failed

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>