Use the mailing list instead of the the private contact address.
Automake has support for test suites, and while trying to use that I ended
up with failing a test and it printed the following message:
    ======================================= [...]
    See tools/test-suite.log
    Please report to replicant@osuosl.org
    ======================================= [...]

As the contact address is only used for matters that cannot be made public,
it would be better to have the bug reported on the mailing list instead,
if testing is ever added in the future.

Other uses of the autotools might also make the private contact address
comes forward pushing users to write to this address for matters that
should be public instead.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>