Initial support for herolte (Samsung Galaxy S7 GSM).
A previous version of this patch was tested on the herolte,
however since then, several light functional changes were
introduced.

With the previous patch, it was possible to boot the modem
and it was probably possible to send message to, and receive
messages from the modem.

For the TARGET_DEVICE we use in our Android.mk, In the
Android.mk of android_device_samsung_herolte[1] we have:
    ifneq ($(filter herolte, $(TARGET_DEVICE)),)

    include $(call all-makefiles-under,$(LOCAL_PATH))

    endif
so we can safely assume that the TARGET_DEVICE is herolte.

[1]https://github.com/LineageOS/android_device_samsung_herolte

Signed-off-by: Tony Garnock-Jones <tonyg@leastfixedpoint.com>
GNUtoo: rebased, code cleanup, more debug prints,
        commit message content but not its summary.
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>