samsung-ipc/utils: file_data_write: return more precise errors
file_data_write uses open and lseek, and both can fail. In that case, it
would be a good idea to be able for the caller of file_data_write to be
able to retrieve the cause of the error.

To do that here, we used the same way than write uses to pass on the
information to the caller:
    On success, the number of bytes written is returned.  On error, -1 is
    returned, and errno is set to indicate the cause of the error.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>