tools: import preliminary rfs-imei implementation as-is
Knowing the format and offset of the IMEI location in nv_data.bin is a
step forward to be able to to recreate nv_data.bin from scratch.
This can be used to repair devices that have their EFS partition
completely destroyed, without having to rely on nonfree tools to do it.

When asking Paul Kocialkowsk if he knew the offset at which the IMEI
were stored in the nv_data.bin files, he explained to me that he only
found it for the XMM616 and also gave me an unfinished implementation
for a tool to change it.

In order to preserve all the information in the tool and its correct
authorship, the tool source code wasn't altered. As it doesn't
compile with current versions of libsamsung-ipc, it wasn't added to
any of the build systems (Android.mk and autotools). This will have
to be fixed in subsequent commits.

Signed-off-by: Paul Kocialkowsk <contact@paulk.fr>
GNUtoo: Added commit message
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>