nv_data-imei: fix leftover from bruteforce_imei_offset
At first I tried to do a bruteforce that computed the IMEI once
for each location in the file. While this was inefficient, it kept
the code simple enough to validate that it worked as the code wasn't
tested before.

However, during the rewrite of that loop to make it more efficient I
forgot to complete it.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>