nv_data-imei: indicate that write-imei may or may not work
I had two Nexus S: after breaking one, I've tried moving the
IMEI of the broken one on the working one:
- Earlier I did some backups of the EFS.
- I went into recovery mode
- I removed the .nv_data.bak and .nv_data.bak.md5 files
- I copied the .nv_data.bin to my computer and changed the
  IMEI with nv_data-imei
- I copied back the .nv_data.bin and the generated .nv_data.bin.md5
  and fixed the permissions.
- I booted the device

However not only the IMEI didn't change, but it was also set back
to the old one in the file.

I've also asked Paul Kocialkowski if he remembered if changing the
IMEI worked in his original code or on which device he tried it,
but he didn't remember.

So more information is probably needed to either understand if
that code is supposed to work and if so if there are conditions
in which it works or doesn't work (specific EFS, devices, etc).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>