Add guix.scm
The guix.scm file can be used to test the build in various
configurations in a reproducible way.

Once you have guix installed, and up to date, you can use
it with the following command:
$ guix build --file=guix.scm

The filename (guix.scm) is a convention that is also used in
other packages to also do build tests in various configurations.

For more background on the use of guix.scm see the Guix days 2020
presentation about it[1].

[1]https://xana.lepiller.eu/guix-days-2020/guix-days-2020-efraim-flashner-build-it-with-guix.webm

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>