Move guix.scm to scripts
The scripts directory already has several scripts that are not
needed to use libsamsung-ipc, but all of them are (or were) useful
to work on patches for libsamsung-ipc.

As this also applies to the guix.scm and that scheme is a scripting
language, it would be more clear to have it moved in the scripts
directory.

In addition it could also make it more simple for distributions
to not ship these files if they aren't interested in having them
and/or if they only want files compatibles with the license
currently used by libsamsung-ipc (GPLv2 or later).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>