Android.mk: Add support for Android 11
If we configure Replicant 11, libsamsung-ipc and libsamsung-ril
to have both libraries in /system/lib/, tools work fine on the
Galaxy SIII (GT-I9300):
    # find -name libsamsung-ril.so 2>/dev/null
    ./system/lib/libsamsung-ril.so

    # find -name libsamsung-ipc.so 2>/dev/null
    ./system/lib/libsamsung-ipc.so

    # find -name nv_data-md5 2>/dev/null
    ./system/bin/nv_data-md5

    # nv_data-md5
    Usage: nv_data-md5 [nv_data.bin]

And libsamsung-ril tries to load libsamsung-ipc from the right
location in /system/lib/, but it fails due to the vendor and system
separation:
    01-14 15:50:57.739  1475  1475 E RILD    : dlopen failed:
    dlopen failed: library "/system/lib/libsamsung-ril.so"
    needed or dlopened by "/system/vendor/bin/hw/rild" is not
    accessible for the namespace "(default)"

Adding LOCAL_PROPRIETARY_MODULE fixes that. Note that the name
of that property can be misleading here: libsamsung-ipc is free
software and shall remain free software.

Instead we need to understand LOCAL_PROPRIETARY_MODULE as a way
to tell the Android build system that libsamsung-ipc is not part
of the base Android code but instead that it is code that is
specific to a device, set of devices and/or Android distribution.

LOCAL_MODULE_RELATIVE_PATH cannot be used instead: even if the
binaies end up in /vendor/bin/hw/, and that the libraries ends
up in /vendor/lib/hw/, for some reasons the libraries can't be
found:
    i9300:/ # /system/vendor/bin/hw/nv_data-md5
    CANNOT LINK EXECUTABLE "/system/vendor/bin/hw/nv_data-md5":
    library "libsamsung-ipc.so" not found: needed by main executable

And most importantly, rild has the same issue than before:
    E RILD    : dlopen failed: dlopen failed: library
                "libsamsung-ipc.so" not found: needed by
                /system/vendor/lib/hw/libsamsung-ril.so
                 in namespace (default)

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>