Add sysexits.h for Android
The include has been imported from sysexits.h from Parabola i686.

As Replicant 6 and Replicant 11 don't have sysexits.h, the
compilation of nv_data-imei will fail if it's not provided somehow:
    hardware/replicant/libsamsung-ipc/tools/nv_data-imei.c:32:22:
    fatal error: sysexits.h: No such file or directory
     #include <sysexits.h>
                      ^
    compilation terminated.

Being able to use sysexits.h is really interesting as it provides
a standard for applications exit codes.

In turn it enables to design tests that are more fine grained
and to avoid too messy test code while doing that, simply by
following this standard.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>