samsung-ipc/utils: add data_read and data_write
These wrappers are meant to handle the case where read or writes
handles a smaller number of bytes than requested.

This way that handling doesn't need to be duplicated everywhere
in libsamsung-ipc.

Having access to the ipc_client struct could enable future
logging of read and writes without having to change any of the
code already using data_read and data_write.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>