MAINTAINERS: Update git URL
While libsamsung-ipc is not limited to Replicant (it can be used
on GNU/Linux or in other Android distributions), it is currently
maintained by Replicant.

In order to make that more clear, it was renamed by the commit
aa4f61805b2d8943e94847cae825ec209060569e ("Move libsamsung-ril
and libsamsung-ipc in vendor/replicant") in the Replicant manifest
repository[1].

Because of that we also need to adjust the git URL in the MAINTAINERS
file.

In addition we use https because it's safer than git:// as it uses
TLS, and because we have redirects in place for http(s) but not for
git://, so if the repository URL changes again, the old URL will
most likely continue to work with http(s).

[1]https://git.replicant.us/replicant/manifest.git

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>