tools: ipc-modem: add option to call a number
The ipc-modem tool contained commented code to call a specific
phone number.

This patch reuses that code and retrieve the number to call from
the command line.

The number to call is expected to be in the same country as the
caller SIM card and network. Adding support for international
calls could be done by computing the right prefix instead of
hardcoding the IPC_CALL_PREFIX_NONE prefix.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>