tools: fix nv-data-md5 compilation with Replicant 6
When compiling nv-data-md5 under Replicant 6 we have:
    hardware/replicant/libsamsung-ipc/tools/nv_data-md5.c:
    In function 'main':
    hardware/replicant/libsamsung-ipc/tools/nv_data-md5.c:103:1:
    error:
    control reaches end of non-void function [-Werror=return-type]
     }
     ^
    cc1: some warnings being treated as errors

So we simply need to add a return in the error path.

While we're at it, we use an exit code referenced in sysexits.h
to follow a standard. This can help differentiating between
different types of errors.

Thanks to that, it can also simplify the code when writing tests.

As Android doesn't have sysexits.h, we also need to use the one that
has been copied inside libsamsung-ipc in tools/include/glibc, else
it would also fail to compile because of the missing header file.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>